# docs/swagger.yaml
basePath: /
definitions:
  headless-cms_internal_models.Article:
    properties:
      content_markdown:
        type: string
      created_at:
        type: string
      id:
        type: integer
      section_id:
        type: integer
      title:
        type: string
      updated_at:
        type: string
    type: object
...
  headless-cms_internal_models.Role:
    enum:
    - user
    - moderator
    - admin
    type: string
    x-enum-varnames:
    - RoleUser
    - RoleModerator
    - RoleAdmin
...
  headless-cms_internal_models.User:
    properties:
      created_at:
        type: string
      id:
        type: integer
      nickname:
        type: string
      role:
        $ref: '#/definitions/headless-cms_internal_models.Role'
      updated_at:
        type: string
    type: object
...
  internal_handler.AuthLoginRequest:
    properties:
      nickname:
        example: alekc
        type: string
      password:
        example: qwerty123
        type: string
    type: object
  internal_handler.AuthLoginResponse:
    properties:
      token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      user:
        $ref: '#/definitions/headless-cms_internal_models.User'
    type: object
  internal_handler.AuthRegisterRequest:
    properties:
      nickname:
        example: alekc
        type: string
      password:
        example: qwerty123
        type: string
    type: object
  internal_handler.ErrorResponse:
    properties:
      error:
        example: описание ошибки
        type: string
    type: object
...
info:
  contact: {}
  description: Headless CMS без фронтенда. Разделы, статьи (Markdown), новости (Markdown
    + автогенерация краткого описания), пользователи/роли, медиа (MinIO).
  title: Headless CMS API
  version: "1.0"
paths:
  /api/v1/articles:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/headless-cms_internal_models.Article'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      summary: Получить все статьи
      tags:
      - Статьи
    post:
      consumes:
      - application/json
      parameters:
      - description: Статья (id игнорируется)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/headless-cms_internal_models.Article'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/headless-cms_internal_models.Article'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Создать статью
      tags:
      - Статьи
  /api/v1/articles/{id}:
    delete:
      parameters:
      - description: ID статьи
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Удалить статью
      tags:
      - Статьи
    get:
      parameters:
      - description: ID статьи
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/headless-cms_internal_models.Article'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      summary: Получить статью по ID
      tags:
      - Статьи
    patch:
      consumes:
      - application/json
      parameters:
      - description: ID статьи
        in: path
        name: id
        required: true
        type: integer
      - description: Статья (id из path)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/headless-cms_internal_models.Article'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/headless-cms_internal_models.Article'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Обновить статью
      tags:
      - Статьи
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: 'Возвращает JWT. Для защищённых методов используйте заголовок Authorization:
        Bearer <token>.'
      parameters:
      - description: Данные входа
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.AuthLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.AuthLoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      summary: Вход и получение JWT токена
      tags:
      - Авторизация
  /api/v1/auth/register:
    post:
      consumes:
      - application/json
      description: Создаёт пользователя с ролью user (обычный читатель).
      parameters:
      - description: Данные регистрации
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.AuthRegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/headless-cms_internal_models.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      summary: Регистрация пользователя
      tags:
      - Авторизация
...
  /api/v1/sections/{id}/articles:
    get:
      parameters:
      - description: ID раздела
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/headless-cms_internal_models.Article'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      summary: Получить статьи раздела
      tags:
      - Статьи
...
securityDefinitions:
  BearerAuth:
    description: 'Введите JWT токен в формате: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"